Rename rustpython_common::{cell -> lock}
