+++ common/src/lib.rs
@@ -1,8 +1,8 @@
 //! A crate to hold types and functions common to all rustpython components.
 
 pub mod borrow;
-pub mod cell;
 pub mod float_ops;
 pub mod hash;
+pub mod lock;
 pub mod rc;
 pub mod str;
