+++ vm/src/stdlib/collections.rs
@@ -2,7 +2,7 @@ pub(crate) use _collections::make_module;
 
 #[pymodule]
 mod _collections {
-    use crate::common::cell::{PyRwLock, PyRwLockReadGuard, PyRwLockWriteGuard};
+    use crate::common::lock::{PyRwLock, PyRwLockReadGuard, PyRwLockWriteGuard};
     use crate::function::OptionalArg;
     use crate::obj::{objiter, objsequence, objtype::PyTypeRef};
     use crate::pyobject::{
