+++ vm/src/stdlib/os.rs
@@ -11,7 +11,7 @@ use num_traits::ToPrimitive;
 
 use super::errno::errors;
 use crate::byteslike::PyBytesLike;
-use crate::common::cell::PyRwLock;
+use crate::common::lock::PyRwLock;
 use crate::exceptions::{IntoPyException, PyBaseExceptionRef};
 use crate::function::{IntoPyNativeFunc, OptionalArg, PyFuncArgs};
 use crate::obj::objbytes::{PyBytes, PyBytesRef};
